# dims-web/src/main/resources/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "更多Spring Boot相关文章请关注：http://blog.didispace.com/",
    "version": "1.0",
    "title": "Spring Boot中使用Swagger2构建RESTful APIs",
    "termsOfService": "http://blog.didispace.com/"
  },
  "host": "localhost:10001",
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/metadata/attributetype": {
      "get": {
        "tags": [
          "metadata-rest-controller"
        ],
        "summary": "attributeType",
        "operationId": "attributeTypeUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "entityTypeId",
            "in": "query",
            "description": "entityTypeId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/metadata/dictionary": {
      "get": {
        "tags": [
          "metadata-rest-controller"
        ],
        "summary": "dictionaries",
        "operationId": "dictionariesUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "dictionaryTypeId",
            "in": "query",
            "description": "dictionaryTypeId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/metadata/entitytype": {
      "get": {
        "tags": [
          "metadata-rest-controller"
        ],
        "summary": "entityType",
        "operationId": "entityTypeUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "entityTypeId",
            "in": "query",
            "description": "entityTypeId",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/metadata/hierarchy": {
      "get": {
        "tags": [
          "metadata-rest-controller"
        ],
        "summary": "hierarchy",
        "operationId": "hierarchyUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "entityTypeId",
            "in": "query",
            "description": "entityTypeId",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "type",
            "in": "query",
            "description": "type",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "ResponseDto": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "object"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "ResponseDto"
    }
  }
}